LayoutTests/media/video-error-does-not-exist-expected.txt

Test that the media element is in correct state after load fails.


EXPECTED (video.error == 'null') OK

EVENT(error)
EXPECTED (video.error != 'null') OK
EXPECTED (video.error.code == '4') OK
EXPECTED (video.networkState == '3') OK
EXPECTED (isNaN(video.duration) == 'true') OK
EXPECTED (video.currentTime == '0') OK
EXPECTED (video.buffered.length == '0') OK
EXPECTED (video.seekable.length == '0') OK
EXPECTED (video.buffered.length == '0') OK
END OF TEST
